+++ Sources/KeyboardKitSwiftUI/Extensions/Color+Button.swift
@@ -28,11 +28,17 @@ import SwiftUI
 public extension Color {
     
     /**
+     The standard light background color in a system keyboard.
+     */
+    static func standardButton(for context: KeyboardContext) -> Color {
+        context.keyboardAppearance == .dark ? .standardDarkAppearanceButton : .standardButton
+    }
+    
+    /**
      The standard button tint color in a system keyboard.
      */
-    static func standardButtonForegroundColor(for context: KeyboardContext) -> Color {
-        if context.userInterfaceStyle == .light && context.keyboardAppearance == .dark { return .white }
-        return .primary
+    static func standardButtonTint(for context: KeyboardContext) -> Color {
+        context.keyboardAppearance == .dark ? .standardDarkAppearanceButtonTint : .standardButtonTint
     }
     
     /**
@@ -45,59 +51,7 @@ public extension Color {
     /**
      The standard dark background color in a system keyboard.
      */
-    static func standardDarkButtonBackgroundColor(for context: KeyboardContext) -> Color {
-        if context.userInterfaceStyle == .dark { return standardDarkButtonBackgroundColorForDarkColorScheme }
-        if context.keyboardAppearance == .dark { return standardDarkButtonBackgroundColorForLightColorSchemeAndDarkKeyboardAppearance }
-        return standardDarkButtonBackgroundColorForLightColorSchemeAndLightKeyboardAppearance
-    }
-    
-    /**
-     The standard light background color in a system keyboard.
-     */
-    static func standardLightButtonBackgroundColor(for context: KeyboardContext) -> Color {
-        if context.userInterfaceStyle == .dark { return standardLightButtonBackgroundColorForDarkColorScheme }
-        if context.keyboardAppearance == .dark { return standardLightButtonBackgroundColorForLightColorSchemeAndDarkKeyboardAppearance }
-        return standardLightButtonBackgroundColorForLightColorSchemeAndLightKeyboardAppearance
-    }
-}
-
-
-// MARK: - Internal Extensions, used for testing
-
-internal extension Color {
-    
-    init(rgb: Double) {
-        self.init(red: rgb, green: rgb, blue: rgb)
-    }
-    
-    static var standardDarkButtonBackgroundColorForDarkColorScheme: Color {
-        Color(rgb: 71.0/255.0)
-    }
-    
-    static var standardDarkButtonBackgroundColorForLightColorSchemeAndDarkKeyboardAppearance: Color {
-        Color(rgb: 117.0/255.0)
-    }
-    
-    static var standardDarkButtonBackgroundColorForLightColorSchemeAndLightKeyboardAppearance: Color {
-        Color(red: 179.0/255.0, green: 184.0/255.0, blue: 193.0/255.0)
-    }
-    
-    static var standardLightButtonBackgroundColorForDarkColorScheme: Color {
-        Color(rgb: 107.0/255.0)
-    }
-    
-    static var standardLightButtonBackgroundColorForLightColorSchemeAndDarkKeyboardAppearance: Color {
-        Color(rgb: 150.0/255.0)
-    }
-    
-    static var standardLightButtonBackgroundColorForLightColorSchemeAndLightKeyboardAppearance: Color {
-        .white
-    }
-}
-
-private extension KeyboardContext {
-    
-    var userInterfaceStyle: UIUserInterfaceStyle {
-        traitCollection.userInterfaceStyle
+    static func standardDarkButton(for context: KeyboardContext) -> Color {
+        context.keyboardAppearance == .dark ? .standardDarkAppearanceDarkButton : .standardDarkButton
     }
 }