+++ Sources/KeyboardKitSwiftUI/_Deprecations/Color+Deprecations.swift
@@ -3,23 +3,15 @@ import SwiftUI
 
 public extension Color {
     
-    @available(*, deprecated, renamed: "standardDarkButtonBackgroundColor")
-    static func systemKeyboardButtonBackgroundColorDark(for context: KeyboardContext) -> Color {
-        standardDarkButtonBackgroundColor(for: context)
-    }
+    @available(*, deprecated, renamed: "standardDarkButton")
+    static func systemKeyboardButtonBackgroundColorDark(for context: KeyboardContext) -> Color { standardDarkButton(for: context) }
     
-    @available(*, deprecated, renamed: "standardLightButtonBackgroundColor")
-    static func systemKeyboardButtonBackgroundColorLight(for context: KeyboardContext) -> Color {
-        standardLightButtonBackgroundColor(for: context)
-    }
+    @available(*, deprecated, renamed: "standardButton")
+    static func systemKeyboardButtonBackgroundColorLight(for context: KeyboardContext) -> Color { standardButton(for: context) }
     
-    @available(*, deprecated, renamed: "standardButtonForegroundColor")
-    static func systemKeyboardButtonForegroundColor(for context: KeyboardContext) -> Color {
-        standardButtonForegroundColor(for: context)
-    }
+    @available(*, deprecated, renamed: "standardButtonTint")
+    static func systemKeyboardButtonForegroundColor(for context: KeyboardContext) -> Color { standardButtonTint(for: context) }
     
     @available(*, deprecated, renamed: "standardButtonShadowColor")
-    static func systemKeyboardButtonShadowColor(for context: KeyboardContext) -> Color {
-        standardButtonShadowColor(for: context)
-    }
+    static func systemKeyboardButtonShadowColor(for context: KeyboardContext) -> Color { standardButtonShadowColor(for: context) }
 }