+++ RELEASE_NOTES.md
@@ -12,17 +12,20 @@ In `v4.0`, this library will become a part of the main library. After that, semv
 * `AutocompleteToolbar` has a new, static  `standardReplacement` function.
 * `AutocompleteToolbar` has a new, static  `standardReplacementAction` function.
 * `AutocompleteToolbar` now lets you provide an optional, custom  `replacementAction` in init.
+* `Color+Resources` is a new extension that provides asset-based colors that adapt to dark mode.
 * `EmojiCategoryKeyboard` is a new view that lists the emojis of a selected category and a menu.
 * `EmojiCategoryKeyboardMenu` is a new view that lets the user select one of multiple categories.
 * `EmojiKeyboard` is a new view that renders a set of emojis in a lazy grid. The item action is customizable.
 
 ### Behavior changes
 
+* `Color+Button` uses the new asset-based colors.
 * `SystemKeyboardButtonBody` now only offsets small caps texts.
 * Due to a secodnary callout bug, the secondary context is now created by the vc instead of the observable keyboard context.
 
 ### Deprecations
 
+* Some button-specific `Color` extensions have been deprecated.
 * `View+keyboardAction(:context:)` has been deprecated.